match history stuff, more consistent entity tracking
