+++ package.json
@@ -145,7 +145,6 @@
 		"zod": "^3.24.1",
 		"zod-to-json-schema": "^3.23.5",
 		"zustand": "^5.0.3",
-		"zustand-rx": "^2.1.0",
 		"zx": "^8.3.2"
 	},
 	"devDependencies": {
