+++ src/components/constraint-violation-display.tsx
@@ -11,20 +11,43 @@ export type ConstraintViolatioDisplayProps = {
 
 export function ConstraintViolationDisplay(props: ConstraintViolatioDisplayProps) {
 	if (props.violated.length == 0) return null
+	const dnrViolations = props.violated.filter(v => v.type === 'do-not-repeat')
+	const filterViolations = props.violated.filter(v => v.type !== 'do-not-repeat')
 	return (
-		<Tooltip>
-			{props.children
-				? <TooltipTrigger asChild>{props.children}</TooltipTrigger>
-				: (
-					<TooltipTrigger>
-						<Icons.ShieldQuestion className="text-orange-400" />
-					</TooltipTrigger>
-				)}
-			<TooltipContent>
-				<ul className="flex flex-col">
-					{props.violated.map(v => <li key={v.id}>{v.name ?? v.id}</li>)}
-				</ul>
-			</TooltipContent>
-		</Tooltip>
+		<>
+			{dnrViolations.length > 0 && (
+				<Tooltip delayDuration={0}>
+					{props.children
+						? <TooltipTrigger asChild>{props.children}</TooltipTrigger>
+						: (
+							<TooltipTrigger>
+								<Icons.ShieldQuestion className="text-pink-400" />
+							</TooltipTrigger>
+						)}
+					<TooltipContent>
+						<div className="font-semibold">DNR:</div>
+						<ul className="flex flex-col">
+							{props.violated.map(v => <li key={v.id}>{v.name ?? v.id}</li>)}
+						</ul>
+					</TooltipContent>
+				</Tooltip>
+			)}
+			{filterViolations.length > 0 && (
+				<Tooltip delayDuration={0}>
+					{props.children
+						? <TooltipTrigger asChild>{props.children}</TooltipTrigger>
+						: (
+							<TooltipTrigger>
+								<Icons.ShieldQuestion className="text-orange-400" />
+							</TooltipTrigger>
+						)}
+					<TooltipContent>
+						<ul className="flex flex-col">
+							{props.violated.map(v => <li key={v.id}>{v.name ?? v.id}</li>)}
+						</ul>
+					</TooltipContent>
+				</Tooltip>
+			)}
+		</>
 	)
 }
