+++ src/components/filter-new.tsx
@@ -10,8 +10,8 @@ import { useToast } from '@/hooks/use-toast'
 import * as EFB from '@/lib/editable-filter-builders.ts'
 import * as M from '@/models.ts'
 
-import { useFilterCreate } from '@/hooks/filters'
 import { assertNever } from '@/lib/typeGuards'
+import { useFilterCreate } from '@/systems.client/filter-entity.client.ts'
 import { invalidateLoggedInUser } from '@/systems.client/logged-in-user'
 import React from 'react'
 import FilterCard from './filter-card'
