+++ src/server/systems/squad-server.ts
@@ -1,6 +1,5 @@
 import * as SchemaModels from '$root/drizzle/schema.models.ts'
-import * as Schema from '$root/drizzle/schema.ts'
-import { AsyncResource, distinctDeepEquals, durableSub, sleep, toAsyncGenerator } from '@/lib/async'
+import { AsyncResource, distinctDeepEquals, toAsyncGenerator } from '@/lib/async'
 import Rcon from '@/lib/rcon/core-rcon.ts'
 import fetchAdminLists from '@/lib/rcon/fetch-admin-lists'
 import * as SM from '@/lib/rcon/squad-models'
@@ -8,7 +7,6 @@ import * as SME from '@/lib/rcon/squad-models.events.ts'
 import SquadRcon from '@/lib/rcon/squad-rcon.ts'
 import { SquadEventEmitter } from '@/lib/squad-log-parser/squad-event-emitter.ts'
 import { assertNever } from '@/lib/typeGuards.ts'
-import { Deferred } from '@/lib/types.ts'
 import { WARNS } from '@/messages.ts'
 import * as M from '@/models.ts'
 import * as RBAC from '@/rbac.models'
@@ -18,25 +16,23 @@ import * as C from '@/server/context.ts'
 import * as DB from '@/server/db.ts'
 import { baseLogger } from '@/server/logger'
 import * as LayerQueue from '@/server/systems/layer-queue.ts'
+import * as MatchHistory from '@/server/systems/match-history.ts'
 import * as Rbac from '@/server/systems/rbac.system.ts'
 import * as Otel from '@opentelemetry/api'
-import { sql } from 'drizzle-orm'
 import * as E from 'drizzle-orm/expressions'
-import { getTableConfig } from 'drizzle-orm/mysql-core'
-import * as Rx from 'rxjs'
 import StringComparison from 'string-comparison'
-import { z } from 'zod'
 import { ENV } from '../env'
 import { procedure, router } from '../trpc.server.ts'
 
 type SquadServerState = {
-	currentMatch?: SM.MatchDetails
-	// stored alongside currentMatch for purposes of reconciliation between rcon and squad events(which are from the server logs), don't use as source of truth generically
-	currentMatchLayerId?: string
+	currentMatchId?: number
 
-	bufferedNewGameLayerSource?: M.LayerSource
-	// the expected entry id for the next game when the layer source was buffered
-	bufferedNewGameEntryId?: number
+	// An ephemeral copy of the current matches layer queue item, taken just before the server rolls and the NEW_GAME squad event is emitted. in order for usages of this to be valid, the corresponding NEW_GAME event must be received after its details are buffered.
+	bufferedNextMatch?: {
+		layerListItem: M.LayerListItem
+		// the expected entry id for the next game when the layer source was buffered
+		historyEntryCount: number
+	}
 }
 
 const tracer = Otel.trace.getTracer('squad-server')
@@ -45,7 +41,7 @@ export let rcon!: SquadRcon
 export let squadLogEvents!: SquadEventEmitter
 
 export let adminList!: AsyncResource<SM.SquadAdmins>
-export let serverStatus!: AsyncResource<SM.ServerStatusResWithCurrentMatch>
+export let serverStatus!: AsyncResource<SM.ServerStatusWithCurrentMatchRes>
 
 export let state!: SquadServerState
 
@@ -228,6 +224,14 @@ export const setupSquadServer = C.spanOp('squad-server:setup', { tracer }, async
 		password: ENV.RCON_PASSWORD,
 	})
 	void coreRcon.connect(ctx)
+		.then(async () => {
+			const statusRes = await rcon.serverStatus.get(ctx)
+			if (statusRes.value.code === 'err:rcon') return
+			const match = MatchHistory.state.recentMatches[0]
+			if (match && M.areLayerIdsCompatible(match.layerId, statusRes.value.data.currentLayer.id)) {
+				state.currentMatchId = match.historyEntryId
+			}
+		})
 	rcon = new SquadRcon(ctx, coreRcon)
 
 	rcon.event$.subscribe(async (event) => {
@@ -253,57 +257,72 @@ export const setupSquadServer = C.spanOp('squad-server:setup', { tracer }, async
 	})
 
 	squadLogEvents.event$.pipe(
-		durableSub('squad-server:handle-squadjs-event', { tracer, ctx }, ([ctx, event]) => handleSquadEvent(DB.addPooledDb(ctx), event)),
+		C.durableSub('squad-server:handle-squadjs-event', { tracer, ctx }, ([ctx, event]) => handleSquadEvent(DB.addPooledDb(ctx), event)),
 	).subscribe()
 	void squadLogEvents.connect()
 
-	serverStatus = new AsyncResource('serverStatusWithCurrentMatch', async (ctx): Promise<SM.ServerStatusResWithCurrentMatch> => {
+	serverStatus = new AsyncResource('serverStatusWithCurrentMatch', async (ctx): Promise<SM.ServerStatusWithCurrentMatchRes> => {
 		const { value: statusRes } = await rcon.serverStatus.get(ctx, {
 			ttl: Math.max(ctx.resOpts.ttl, 5000 / 2),
 		})
 		if (statusRes.code !== 'ok') return statusRes
-		const res: SM.ServerStatusResWithCurrentMatch = { code: 'ok' as const, data: { ...statusRes.data } }
-		if (state.currentMatchLayerId && M.areLayerIdsCompatible(statusRes.data.currentLayer.id, state.currentMatchLayerId)) {
-			res.data.currentMatch = state.currentMatch!
+		const res: SM.ServerStatusWithCurrentMatchRes = { code: 'ok' as const, data: { ...statusRes.data } }
+		const match = MatchHistory.state.recentMatches[0]
+		if (match && match?.historyEntryId === state.currentMatchId && M.areLayerIdsCompatible(match.layerId, statusRes.data.currentLayer.id)) {
+			res.data.currentMatchId = state.currentMatchId
 		}
 		return res
 	}, { defaultTTL: 5000 })
 
-	void getCurrentMatchDetailsFromDb(ctx).then(async (res) => {
-		if (!res) return
-		const [currentMatch, layerId] = res
-		state.currentMatch = currentMatch
-		state.currentMatchLayerId = layerId
-		serverStatus.invalidate(ctx)
-	})
 	C.getSpan()!.setStatus({ code: Otel.SpanStatusCode.OK })
 })
 
 async function handleSquadEvent(ctx: C.Log & C.Db, event: SME.Event) {
 	switch (event.type) {
 		case 'NEW_GAME': {
-			const { value: statusRes } = await rcon.serverStatus.get(ctx, { ttl: 200 })
-			if (statusRes.code !== 'ok') return statusRes
-			const entry: SchemaModels.NewMatchHistory = {
-				layerId: statusRes.data.currentLayer.id,
-				startTime: event.time,
-				setByType: state.bufferedNewGameLayerSource?.type ?? 'unknown',
-				// setByUserId: state.bufferedNewGameLayerSource?.type ==
-			}
-			await ctx.db().insert(Schema.matchHistory).values(entry)
-			return { code: 'ok' as const }
+			return await DB.runTransaction(ctx, async (ctx) => {
+				const { value: statusRes } = await rcon.serverStatus.get(ctx, { ttl: 200 })
+				if (statusRes.code !== 'ok') return statusRes
+				let newEntry: SchemaModels.NewMatchHistory = {
+					layerId: statusRes.data.currentLayer.id,
+					startTime: event.time,
+					setByType: 'unknown',
+					// setByUserId: state.bufferedNewGameLayerSource?.type ==
+				}
+				if (state.bufferedNextMatch) {
+					const setByUserId = state.bufferedNextMatch.layerListItem.source.type === 'manual'
+						? state.bufferedNextMatch.layerListItem.source.userId
+						: undefined
+					const updated: Partial<SchemaModels.NewMatchHistory> = {
+						setByType: state.bufferedNextMatch.layerListItem.source.type,
+						setByUserId: setByUserId,
+					}
+					newEntry = {
+						...newEntry,
+						...updated,
+					}
+				}
+
+				const res = await MatchHistory.addHistoryEntry(ctx, newEntry, state.bufferedNextMatch?.historyEntryCount)
+				if (res.code !== 'ok') return res
+				state.currentMatchId = res.match.historyEntryId
+
+				delete state.bufferedNextMatch
+
+				return { code: 'ok' as const }
+			})
 		}
+
 		case 'ROUND_ENDED': {
 			const { value: statusRes } = await rcon.serverStatus.get(ctx, { ttl: 200 })
 			if (statusRes.code !== 'ok') return statusRes
 
 			await DB.runTransaction(ctx, async (ctx) => {
-				const [currentMatch] = await ctx.db().select()
-					.from(Schema.matchHistory)
-					.orderBy(E.desc(Schema.matchHistory.startTime))
-					.limit(1)
-					.for('update')
-				if (!currentMatch || !M.areLayerIdsCompatible(currentMatch.layerId, statusRes.data.currentLayer.id)) return
+				const currentMatch = MatchHistory.state.recentMatches[0]
+				if (!currentMatch || !M.areLayerIdsCompatible(currentMatch.layerId, statusRes.data.currentLayer.id)) {
+					delete state.currentMatchId
+					return
+				}
 				const teams: [SM.SquadOutcomeTeam | null, SM.SquadOutcomeTeam | null] = [event.winner, event.loser]
 				if (teams[0]) teams.sort((a, b) => a!.team - b!.team)
 				const outcome = event.winner === null ? 'draw' : event.winner.team === 1 ? 'team1' : 'team2'
@@ -311,13 +330,10 @@ async function handleSquadEvent(ctx: C.Log & C.Db, event: SME.Event) {
 					endTime: event.time,
 					team1Tickets: teams[0]?.tickets,
 					team2Tickets: teams[1]?.tickets,
-					layerId: statusRes.data.currentLayer.id,
 					outcome,
 				}
 
-				await ctx.db().update(Schema.matchHistory)
-					.set(entry)
-					.where(E.eq(Schema.matchHistory.id, currentMatch.id))
+				state.currentMatchId = (await MatchHistory.finalizeCurrentHistoryEntry(ctx, entry))?.historyEntryId
 			})
 			return { code: 'ok' as const }
 		}
@@ -326,36 +342,12 @@ async function handleSquadEvent(ctx: C.Log & C.Db, event: SME.Event) {
 	}
 }
 
-export async function getCurrentMatchDetailsFromDb(ctx: C.Log & C.Db): Promise<[SM.MatchDetails, M.LayerId] | null> {
-	const [entry] = await ctx.db().select()
-		.from(Schema.matchHistory)
-		.orderBy(E.desc(Schema.matchHistory.startTime))
-		.limit(1)
-		.for('update')
-
-	if (!entry || entry.endTime !== null) return null
-
-	try {
-		return [SM.historyEntryToMatchDetails(entry), entry.layerId]
-	} catch (error) {
-		console.error(error, 'Error fetching match details:')
-		return null
+export async function bufferNextMatchLQItem(ctx: C.Db & C.Log, item: M.LayerListItem) {
+	const count = await MatchHistory.getMatchHistoryCount(ctx)
+	state.bufferedNextMatch = {
+		historyEntryCount: count,
+		layerListItem: item,
 	}
-}
-
-async function getNextMatchHistoryEntryId(ctx: C.Log & C.Db): Promise<number> {
-	// @ts-expect-error idc
-	const [[row]] = await ctx.db().execute(sql<number>`
-      SELECT AUTO_INCREMENT FROM information_schema.tables
-      WHERE TABLE_SCHEMA = ${ENV.DB_DATABASE} AND TABLE_NAME = ${getTableConfig(Schema.matchHistory).name}
-     `)
-
-	return parseInt(row.AUTO_INCREMENT)
-}
-
-// expose setting the current layer source when the next layer starts
-async function setCurrentLayerSource(source: M.LayerSource) {
-	state.bufferedNewGameLayerSource = source
 }
 
 export const squadServerRouter = router({
